+++ samples/bluetooth/peripheral_ht/CMakeLists.txt
@@ -8,7 +8,4 @@ project(peripheral_ht)
 FILE(GLOB app_sources src/*.c)
 target_sources(app PRIVATE
   ${app_sources}
-  ../gatt/hts.c
   )
-
-zephyr_library_include_directories($ENV{ZEPHYR_BASE}/samples/bluetooth)
