+++ samples/bluetooth/peripheral_ht/src/hts.c
