+++ samples/bluetooth/peripheral_ht/src/hts.h
